replaced provider list with provider cards
