+++ client/src/components/ProviderPage.js
@@ -1,9 +1,18 @@
-import React, { useState } from "react";
-import ProvidersList from "./ProvidersList";
+import React, { Suspense, useState, useEffect } from "react";
+import ProviderCard from "./ProviderCard";
 import ProviderForm from "./ProviderForm";
 
 function ProviderPage() {
   const [providers, setProviders] = useState([]);
+
+  useEffect(() => {
+    const fetchProviders = async () => {
+      const response = await fetch("/providers");
+      const provArr = await response.json();
+      setProviders(provArr);
+    };
+    fetchProviders().catch(console.error);
+  }, []);
 
   const handleNewProvider = (newProvider) => {
     setProviders([...providers, newProvider]);
@@ -19,15 +28,21 @@ function ProviderPage() {
     });
   }
 
+  let providerCards = providers.map((provider) => (
+    <ProviderCard
+      key={provider.id}
+      provider={provider}
+      onDelete={handleDeleteProvider}
+    />
+  ));
+
   return (
     <div>
       <h1>Provider Portal</h1>
-      <ProvidersList
-        providers={providers}
-        setProviders={setProviders}
-        handleNewProvider={handleNewProvider}
-        onDelete={handleDeleteProvider}
-      />
+      <Suspense>
+        <h1>Providers</h1>
+        <div className="providerList">{providerCards}</div>
+      </Suspense>
       <ProviderForm handleNewProvider={handleNewProvider} />
     </div>
   );
